fix issue

+++ go.sh
@@ -4,5 +4,7 @@ if [ $UID -ne 0 ];then
 	exit
 fi 
 
-rm -rf work
+if [ $1 = 'clean' ];then
+	rm -rf work
+fi
 ./build.sh -N archlinux-lizr-edition -P 'Lizr <oilizr2004@gmail.com>' -v
